CDAP-18632: Move into own package to ensure it does not clash with common io plugins; properly skip validation with macros, use classloader when getting filesystem
